--- MANIFEST.in
@@ -1,1 +1,0 @@
-include bitcoin/english.txt
