--- bitcoin/explorers/__init__.py
@@ -1,1 +1,0 @@
-from . import base_insight, bitpay, blockcypher, blockdozer, dash_siampm, sochain
