+++ cryptos/coins/base.py
