+++ cryptos/coins/litecoin.py
