+++ cryptos/english.txt
