+++ cryptos/explorers/blockcypher.py
@@ -1,19 +1,6 @@
 import re
 from blockcypher import api
-
-def parse_addr_args(*args):
-    # Valid input formats: unspent([addr1, addr2, addr3])
-    #                      unspent([addr1, addr2, addr3], coin_symbol="btc")
-    #                      unspent(addr1, addr2, addr3)
-    #                      unspent(addr1, addr2, addr3, coin_symbol="btc")
-    addr_args = args
-    if len(args) == 0:
-        return []
-    if len(addr_args) == 1 and isinstance(addr_args[0], list):
-        addr_args = addr_args[0]
-    elif isinstance(addr_args, tuple):
-        addr_args = list(addr_args)
-    return addr_args
+from .utils import parse_addr_args
 
 def unspent(*addrs, coin_symbol=None, api_key=None, **kwargs):
