+++ cryptos/mnemonic.py
