+++ cryptos/py2specials.py
