+++ cryptos/py3specials.py
