+++ setup.py
@@ -11,7 +11,7 @@ setup(name='cryptos',
       author_email='paulmartinforwork@gmail.com',
       url='http://github.com/primal100/pybitcointools',
       packages=['cryptos'],
-      scripts=['pybtctool'],
+      scripts=['cryptotool'],
       include_package_data=True,
-      data_files=[("", ["LICENSE"]), ("bitcoin", ["bitcoin/english.txt"])],
+      data_files=[("", ["LICENSE"]), ("cryptos", ["cryptos/english.txt"])],
       )
